spanner_agent/requirements.txt
google-adk[a2a]>=1.24.0
google-cloud-spanner>=3.0.0
google-auth>=2.0.0
a2a-sdk>=0.3.0
